Fix error in groupArrayMovingSum
