+++ tests/queries/0_stateless/02817_group_array_moving_zero_window_size.sql
@@ -1,0 +1,2 @@
+SELECT groupArrayMovingAvg ( toInt64 ( 0 ) ) ( toDecimal32 ( 1 , 1 ) ); -- { serverError BAD_ARGUMENTS }
+
